storage/db: read multi-line records from database
The database format allows multiple lines to be stored in a single
record, to reduce the number of rows in the database. This change
allows reading such records, though there is not yet any code which
can write them. This change also adds a test that will benefit from
this record compression.

Change-Id: Ia86153fb150534d701d01a0c102be5a3c29b826d
Reviewed-on: https://go-review.googlesource.com/35672
Reviewed-by: Russ Cox <rsc@golang.org>