storage/db: ListUploads shouldn't return empty uploads
The Uploads table gets an entry when a user or the server aborts an
upload in mid-upload. The Records table is cleaned up by the
transaction, but the Uploads entry is left behind.

The optimized query that is used for the perf.golang.org homepage did
not properly exclude these uploads that have no records.

Change-Id: I8322af5946c233e5cb4b15bf57d9487bcd1e6423
Reviewed-on: https://go-review.googlesource.com/39051
Run-TryBot: Quentin Smith <quentin@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>