benchstat: don't print sign for 0% when there is no change
When all samples are identical, instead of printing
+0.00%, print just 0.00%.

This is not just more accurate.
When benchmarking the compiler, one item of interest
is the impact on object file and executable sizes.
These can change by tiny amounts, below two sig figs.
However, the existence of any change is sometimes a cue
to the reader to investigate further.
This obviously should not be relied on,
but it is nevertheless a useful signal.

Change-Id: I1e5157a52fc47c6871ca87b0e9cff223f4fd43d9
Reviewed-on: https://go-review.googlesource.com/44190
Run-TryBot: Josh Bleecher Snyder <josharian@gmail.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>