benchproc: new package for high-level benchmark processing
This provides a transform/filter/projection processing model for
working with benchmark results, and mini-languages for constructing
filters and projections. This model is the result of quite a bit of
experimentation, and at this point I've built three tools on top of
this package, so I consider it reasonably well validated.

This also introduces a benchproc/syntax package, akin to the
regexp/syntax package. The syntax package documents the syntax and
provides low-level parsers for filter and projection expressions. If
we ever port the web UI over to this, this will make it easy to
transform these expressions into database filters.

This requires updating the language version in go.mod to 1.13 because
benchproc uses binary literals and signed shifts.

Change-Id: If9da95ec1e583594644c123f3bc3fc377a0e97e5
Reviewed-on: https://go-review.googlesource.com/c/perf/+/283617
Trust: Austin Clements <austin@google.com>
Run-TryBot: Austin Clements <austin@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>