benchseries: ignore results with mismatched compare value
Currently, benchseries assumes that results with compare value ==
denominator is in the denominator set, and everything else is in the
numerator.

In addition to being confusing (BuilderOptions.Numerator is completely
ignored), this breaks things when the compare key isn't set at all. In
that case cmpCfg.StringValues() is always "", so everything ends up in
the numerator and nothing in the denominator, which causes a later crash
due to a missing denominator [1].

Be more explicit and only save results that explicitly fall into the
numerator or denominator.

[1] This fix isn't perfect, as benchmark results that contain only
numerator results and no denominator results will still cause a crash.
That should probably be an error or simply result in no comparisons.

For golang/go#61362.

Change-Id: I1444d0fcb101bcea0f1cdf31de7292211dcbc901
Reviewed-on: https://go-review.googlesource.com/c/perf/+/510475
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Michael Pratt <mpratt@google.com>
Run-TryBot: Michael Pratt <mpratt@google.com>
Reviewed-by: David Chase <drchase@google.com>